+++ npm/webpack-batteries-included-preprocessor/index.ts
@@ -191,10 +191,6 @@ const getDefaultWebpackOptions = () => {
             babelrc: false,
           },
         }],
-      }, {
-        test: /\.coffee$/,
-        exclude: [/node_modules/, /browserslist/],
-        loader: require.resolve('coffee-loader'),
       }],
     },
     plugins: [
@@ -222,7 +218,7 @@ const getDefaultWebpackOptions = () => {
       ...(Debug.enabled(WBADebugNamespace) ? [new BundleAnalyzerPlugin()] : []),
     ],
     resolve: {
-      extensions: ['.js', '.json', '.jsx', '.mjs', '.coffee'],
+      extensions: ['.js', '.json', '.jsx', '.mjs'],
       fallback: {
         assert: false,
         buffer: require.resolve('buffer/'),